requirements_mac.txt
matplotlib==3.1.1
numpy==1.17.2
scikit-learn==0.21.3
torchvision
scipy==1.3.1
autograd==1.3
progressbar2==3.43.1
jupyterlab==1.1.3
jupyterlab-server==1.0.6
